@@ -11,46 +11,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.0.7",
-    "@angular/common": "9.0.7",
-    "@angular/compiler": "9.0.7",
-    "@angular/core": "9.0.7",
-    "@angular/forms": "9.0.7",
-    "@angular/platform-browser": "9.0.7",
-    "@angular/platform-browser-dynamic": "9.0.7",
-    "@angular/router": "9.0.7",
-    "core-js": "3.6.4",
-    "rxjs": "6.5.4",
-    "tslib": "1.11.1",
-    "zone.js": "0.10.3",
+    "@angular/animations": "10.2.0",
+    "@angular/common": "10.2.0",
+    "@angular/compiler": "10.2.0",
+    "@angular/core": "10.2.0",
+    "@angular/forms": "10.2.0",
+    "@angular/platform-browser": "10.2.0",
+    "@angular/platform-browser-dynamic": "10.2.0",
+    "@angular/router": "10.2.0",
     "@babylonjs/core": "4.1.0",
-    "@babylonjs/materials": "4.1.0",
+    "@babylonjs/gui": "4.1.0",
+    "@babylonjs/inspector": "4.1.0",
     "@babylonjs/loaders": "4.1.0",
+    "@babylonjs/materials": "4.1.0",
     "@babylonjs/post-processes": "4.1.0",
     "@babylonjs/procedural-textures": "4.1.0",
     "@babylonjs/serializers": "4.1.0",
-    "@babylonjs/gui": "4.1.0",
-    "@babylonjs/inspector": "4.1.0"
+    "@ngx-translate/core": "^13.0.0",
+    "core-js": "3.6.4",
+    "rxjs": "6.5.4",
+    "tslib": "^2.0.0",
+    "zone.js": "0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.1002.0",
-    "@angular/cli": "9.0.7",
-    "@angular/compiler-cli": "9.0.7",
-    "@angular/language-service": "9.0.7",
+    "@angular/cli": "10.2.0",
+    "@angular/compiler-cli": "10.2.0",
+    "@angular/language-service": "10.2.0",
     "@types/jasmine": "3.5.9",
     "@types/jasminewd2": "2.0.8",
     "@types/node": "13.9.2",
     "codelyzer": "5.2.1",
-    "jasmine-core": "3.5.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.4.1",
-    "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-jasmine": "3.1.1",
-    "karma-jasmine-html-reporter": "1.5.2",
-    "protractor": "5.4.3",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
     "ts-node": "8.7.0",
-    "tslint": "6.1.0",
-    "typescript": "~3.7.5"
+    "tslint": "~6.1.0",
+    "typescript": "3.9.7"
   }
 }
