@@ -1,9 +1,9 @@
-import { Component, OnInit } from "@angular/core";
+import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: "app-options-menu",
-  templateUrl: "./options-menu.component.html",
-  styleUrls: ["./options-menu.component.scss"]
+  selector: 'app-options-menu',
+  templateUrl: './options-menu.component.html',
+  styleUrls: ['./options-menu.component.scss']
 })
 export class OptionsMenuComponent implements OnInit {
   activeMenu: string;
